Add files via upload
This contains the rule set and the prompt for the user to draw the rows and columns. I wasn't too sure about the draw() function and whether or not I was supposed to do it or if it's part of Esther's function. So just take a look and let me know.